requirements.txt
# Core ML & Deep Learning
torch>=2.0.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# Causal Inference (UNIQUE!)
dowhy>=0.10.0
econml>=0.14.0
causalml>=0.14.0

# Graph & Network Analysis
networkx>=3.0
torch-geometric>=2.3.0

# Semantic Understanding
sentence-transformers>=2.2.0
transformers>=4.30.0

# Fast Similarity Search
faiss-cpu>=1.7.4

# Gradient Boosting
lightgbm>=4.0.0
xgboost>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
polars>=0.18.0

# Scientific Computing
scipy>=1.10.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# API & Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Caching & Storage
redis>=4.6.0
hiredis>=2.2.0

# Monitoring & Logging
prometheus-client>=0.17.0
python-json-logger>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Experiment Tracking
mlflow>=2.5.0
wandb>=0.15.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0
